api/requirements.txt
flask
flask-cors
python-dotenv
gunicorn
gitpython
redis
rq
prisma
langchain>=0.1.0
langchain-community>=0.0.10
langchain-google-genai
langchain-core
sqlmodel
psycopg2-binary
requests